Change CustomLinkedListElement into inner class inside CustomLinkedList, fix iterator()
